--- src/main/java/ru/practicum/shareit/request/item/controller/ItemController.java
@@ -1,44 +1,0 @@
-package ru.practicum.shareit.request.item.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.request.item.dto.ItemDto;
-import ru.practicum.shareit.request.item.model.Item;
-import ru.practicum.shareit.request.item.service.ItemService;
-
-import javax.validation.Valid;
-import java.util.List;
-
-@RestController
-@RequestMapping("/items")
-@RequiredArgsConstructor(onConstructor_ = @Autowired)
-public class ItemController {
-
-    private final ItemService itemService;
-
-    @PostMapping
-    public ItemDto create(@RequestHeader("X-Sharer-User-Id") long userId, @Valid @RequestBody ItemDto itemDto) {
-        return itemService.create(userId, itemDto);
-    }
-
-    @PatchMapping("/{itemId}")
-    public ItemDto update(@RequestHeader("X-Sharer-User-Id") long userId, @PathVariable long itemId, @RequestBody Item item) {
-        return itemService.update(userId, itemId, item);
-    }
-
-    @GetMapping
-    public List<ItemDto> findAll(@RequestHeader("X-Sharer-User-Id") long id) {
-        return itemService.findAll(id);
-    }
-
-    @GetMapping("/{itemId}")
-    public ItemDto findItem(@PathVariable long itemId) {
-        return itemService.findItem(itemId);
-    }
-
-    @GetMapping("/search")
-    public List<ItemDto> searchItem(@RequestParam String text) {
-        return itemService.searchItem(text);
-    }
-}
